ADDED:test code to compare 3 feet with 1 yard

+++ src/main/java/com/bridglabz/Inch.java
@@ -1,7 +1,8 @@
 package com.bridglabz;
 
+
 public class Inch extends Units {
-    public Inch(int length) {
+    public Inch(double length) {
         super(length);
     }
 }
